Refactored code and replaced Dagger2 with Kodein
Fixed comment reviews
Refactored the folder structure
Removed unused classes
