+++ app/src/main/java/com/example/bankdetails/BanksApplication.kt
@@ -1,10 +1,26 @@
 package com.example.bankdetails
 
 import android.app.Application
-import com.example.bankdetails.injection.component.DaggerAppComponent
+import androidx.room.Room
+import com.example.bankdetails.database.BanksDao
+import com.example.bankdetails.database.BanksDatabase
+import org.kodein.di.*
 
-class BanksApplication : Application() {
-    val appComponent by lazy {
-        DaggerAppComponent.factory().create(this)
+class BanksApplication : Application(), DIAware {
+
+    override val di by DI.lazy {
+
+        bind<BanksDatabase>("database") with singleton {
+            Room.databaseBuilder(
+                applicationContext,
+                BanksDatabase::class.java,
+                "banks_database"
+            ).build()
+        }
+
+        bind<BanksDao>("databaseDao") with singleton {
+            instance<BanksDatabase>("database").banksDao()
+        }
+
     }
 }
