--- app/src/main/java/com/example/bankdetails/data/database/BanksListDao.kt
@@ -1,6 +1,0 @@
-package com.example.bankdetails.data.database
-
-import androidx.room.Dao
-
-@Dao
-interface BanksListDao
