--- app/src/main/java/com/example/bankdetails/data/database/BanksListDatabase.kt
@@ -1,27 +1,0 @@
-package com.example.bankdetails.data.database
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import com.example.bankdetails.data.model.BankInfo
-
-@Database(entities = [BankInfo::class], version = 1, exportSchema = false)
-abstract class BanksListDatabase : RoomDatabase() {
-    abstract fun banksListDao(): BanksListDao
-
-    companion object {
-        @Volatile
-        var INSTANCE: BanksListDatabase? = null
-
-        fun getInstance(context: Context): BanksListDatabase? {
-            if (INSTANCE == null) {
-                synchronized(this) {
-                    Room.databaseBuilder(context, BanksListDatabase::class.java, "banks_database")
-                        .fallbackToDestructiveMigration().build()
-                }
-            }
-            return INSTANCE
-        }
-    }
-}
