+++ app/src/main/java/com/example/bankdetails/database/BanksDao.kt
@@ -1,0 +1,6 @@
+package com.example.bankdetails.database
+
+import androidx.room.Dao
+
+@Dao
+interface BanksDao
