--- app/src/main/java/com/example/bankdetails/injection/scope/ViewModelScope.kt
@@ -1,11 +1,0 @@
-package com.example.bankdetails.injection.scope
-
-import androidx.lifecycle.ViewModel
-import dagger.MapKey
-import kotlin.reflect.KClass
-
-@MustBeDocumented
-@Target(AnnotationTarget.FUNCTION)
-@Retention(AnnotationRetention.RUNTIME)
-@MapKey
-annotation class ViewModelScope(val value: KClass<out ViewModel>)
