+++ app/src/main/java/com/example/bankdetails/model/BankInfo.kt
@@ -1,4 +1,4 @@
-package com.example.bankdetails.data.model
+package com.example.bankdetails.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
