+++ app/src/main/res/layout/item_banks.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="bankData"
-            type="com.example.bankdetails.data.model.BankInfo" />
+            type="com.example.bankdetails.model.BankInfo" />
     </data>
 
     <androidx.cardview.widget.CardView
@@ -24,7 +24,7 @@
             android:padding="@dimen/space_8dp">
 
             <TextView
-                android:id="@+id/bank_label"
+                android:id="@+id/bankLabel"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:fontFamily="sans-serif"
@@ -37,21 +37,21 @@
                 app:layout_constraintTop_toTopOf="parent" />
 
             <TextView
-                android:id="@+id/ifsc_label"
+                android:id="@+id/ifscLabel"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:padding="4dp"
                 android:text="@{ `IFSC: `+ bankData.ifsc}"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/bank_label" />
+                app:layout_constraintTop_toBottomOf="@+id/bankLabel" />
 
             <CheckBox
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:button="@drawable/selector_favorite"
-                app:layout_constraintBaseline_toBaselineOf="@id/bank_label"
+                app:layout_constraintBaseline_toBaselineOf="@id/bankLabel"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintLeft_toRightOf="@id/bank_label" />
+                app:layout_constraintLeft_toRightOf="@id/bankLabel" />
 
 
         </androidx.constraintlayout.widget.ConstraintLayout>
